Add if statement to launch.sh script to check which window manager is being used

+++ .config/polybar/launch.sh
@@ -4,6 +4,13 @@
 # If all your bars have ipc enabled, you can use 
 # polybar-msg cmd quit
 # Otherwise you can use the nuclear option:
+
+#only launch Polybar if not running i3 wm
+if [[ "$XDG_CURRENT_DESKTOP" == "i3" ]] || [[ "$DESKTOP_SESSION" == "i3" ]]; then
+    echo "i3 session detected - skipping Polybar launch."
+    exit 0
+fi
+
 killall -q polybar
 
 # Launch bar1 and bar2
